move run_tests to test_utils
